+++ river_route/uhkernels/__init__.py
@@ -1,6 +1,6 @@
-from .UnitHydrograph import UnitHydrograph
 from .SCSCurvilinear import SCSCurvilinear
 from .SCSTriangular import SCSTriangular
+from .UnitHydrograph import UnitHydrograph
 
 __all__ = [
     'UnitHydrograph',