+++ tests/test_muskingum.py
@@ -22,9 +22,9 @@ def test_muskingum_channel_only(vpu: RFSv2ConfigsData):
         pd.DataFrame({'Q': np.full(n_rivers, 10.0)}).to_parquet(init_state_file)
 
         discharge_file = os.path.join(tmpdir, 'q_channel_only.nc')
-        final_state_file = os.path.join(tmpdir, 'final_state.parquet')
+        # final_state_file = os.path.join(tmpdir, 'final_state.parquet')
 
-        rr.Muskingum(
+        rr.Router(forcing='channel', 
             params_file=str(vpu.rr2_params_file),
             discharge_files=[discharge_file],
             channel_state_init_file=init_state_file,
@@ -58,7 +58,7 @@ def test_muskingum_zero_initial_state(vpu: RFSv2ConfigsData):
 
         discharge_file = os.path.join(tmpdir, 'q.nc')
 
-        rr.Muskingum(
+        rr.Router(forcing='channel', 
             params_file=str(vpu.rr2_params_file),
             discharge_files=[discharge_file],
             channel_state_init_file=init_state_file,